To prevent the startup from cashing out, correctly disables the ports configuratively. You see, the emulator was failing when it was starting up with a KeyErro: 'interface' and that's what we got when we used the 'prod_config.yml'. That is what we got from the en3 and en2 configurations that, they lacked a key interface but weren't even marked properly as inactive.
So, we corrected the configuration so you can uncomment the type: disconnected directive while these ports, exist within the logical construct of the class Sim to skip the initialization thing, which resolves the whole error thing which just shows, what that really means.

It becomes a configurative change which is "so much more than" just a simple fix bug..it demonstrates a principle that is the underlying concept behind the Graph Virtual Machine..by selectively de-activating and activating links within a file configuration, we can alter instead the topology of the computations for yea, the purpose of "altering" analysis; which makes the controlled & precise scenarios of emulation something that we don't have to resort to at least, the last irreversible smash & restart, of information of Shannon which our adaptive system, deletes or failures or i.e. removes, the resources all old and, this method of port management, is a direct designation of that adaptive behavior.

Thus no longer do we have to correct the configuration by uncommenting the type: disconnected directive for these ports..thus no longer do we have to utilize the existing loci logic but we do, the logic in the class of Sim to skip what is their initialization which, re-solves the error.